Added a potential remediation for a problem that couldn't be recreated
A DUPLICATE_FIELD exception was coming back from Jamf Pro, but I wasn't able to duplicate the issue or determine a root cause, however I did determine that this happened when something was added when it should have been updated. When this error is detected, it will attempt to get the package info from the server again and if the package is found, it will update the package instead of adding it. It will show a warning in this case. If it still can't find the package, it will show an error, otherwise it will log a message indicating that it succeeded.