Mostly fixed the issue with the progress bar
There may be some stuff to tweak yet. For instance, if a chunk fails to update, it should re-add the size of that chunk to SynchronizationProgress's totalSize value. Also, it seems like once a chunk is finished, the progress bar falls back for some reason and it should just keep moving forward. I tried cancelling the upload and that didn't seem to work either.