Added the active flag to the Jamf Pro Server view and made sure that DPs from inactive Jamf Pro instances are not included in the DP lists
Also fixed an issue where if a change was made and later it was canceled, it would reload even though it was canceled. Made the same fix with Setup. Also made sure that the keychain data is loaded even if it's not active so it shows up when editing a Jamf Pro server.